server/requirements.txt
asgiref==3.5.2
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==23.0.0
mypy-extensions==0.4.3
packaging==24.2
pathspec==0.10.3
pillow==11.0.0
platformdirs==2.6.0
psycopg2-binary==2.9.10
pytz==2022.6
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2024.2
uritemplate==4.1.1
wheel==0.45.1
whitenoise==6.8.2
